wasm-bindgen: bundle
