+++ packages/web-platform/web-style-transformer/index.js
@@ -1,0 +1,6 @@
+export let wasm;
+export function initWasm() {
+  return import('./standard.js').then((module) => {
+    wasm = module;
+  });
+}
