+++ packages/web-platform/web-style-transformer/package.json
@@ -10,19 +10,20 @@
   "license": "Apache-2.0",
   "sideEffects": false,
   "type": "module",
-  "main": "dist/index.js",
+  "main": "index.js",
   "typings": "index.d.ts",
   "files": [
     "dist",
     "!dist/**/*.js.map",
-    "index.d.ts",
+    "*.d.ts",
+    "*.js",
     "LICENSE.txt",
     "Notice.txt",
     "CHANGELOG.md",
     "README.md"
   ],
   "scripts": {
-    "build": "wasm-pack build --target web --out-dir dist --out-name index --mode no-install && node -e \"require('fs').rmSync('dist/.gitignore', { force: true })\"",
+    "build": "wasm-pack build --target bundler --out-dir dist --out-name index --mode no-install && node -e \"require('fs').rmSync('dist/.gitignore', { force: true })\"",
     "build:wasm-bindgen": "wasm-bindgen --version || cargo install wasm-bindgen-cli",
     "build:wasm-pack": "wasm-pack --version || cargo install wasm-pack"
   }
