+++ packages/web-platform/web-style-transformer/standard.js
@@ -1,0 +1,2 @@
+export * from './dist/index.js';
+export { memory } from './dist/index_bg.wasm';
